Init commit

+++ src/routers/task.js
@@ -1,14 +1,16 @@
 const express = require('express')
 const router = new express.Router()
 const Task = require('../model/tasks')
-const validator = require('validator')
-const User = require('../model/users')
+const auth  = require('../middleware/auth')
 
 
 
 router.post('/tasks', async(req, res) => {
-    const task = new Task(req.body)
-
+    //const task = new Task(req.body)
+         const task = new Task({
+             ...req.body,
+             owner: req.user._id
+         })
 
     try {
         await task.save()
@@ -27,11 +29,43 @@ router.post('/tasks', async(req, res) => {
 })
 
 //---------------------read all data-----------------------------------------
-router.get('/tasks', async(req, res) => {
+//filtering the task GET/ tasks?completed=trueorfasle
+//pagination limit skip GET/taks?limit=3&skip=0
+//sort by GET/tasks?sortBy=createdAt_asc/desc         0to10/10to0
+router.get('/tasks', auth, async(req, res) => {
+      const match = {}
+      const sort = {}
 
+       if(req.query.completed){
+               match.completed = req.query.completed === 'true'
+      }
+    if(req.query.sortBy){
+        const parts = req.query.sortBy.split(':')
+        sort[parts[0]] = parts[1] === 'desc'? -1 : 1        //asc/desc
+    }
      try {
-         const task = await Task.find({})
+        //  const task = await Task.find({})
+         //const task = await Task.find({owner: req.user._id})
+         await req.user.populate({
+            path: 'task',
+            match,
+            options: {
+              limit: parseInt(req.query.limit),
+              skip: parseInt(req.query.skip),
+              sort
+              ////sort: {
+                   // //  createdAt: 1 for ascending
+                   // //  createdAt: -1 for descending
+                   // completed: -1 //for false 1 for true
+              // //}
+
+            }
+                
+            }
+            ).execPopluate()
+         
          res.status(200).send(task)
+         // res.send(req.user.task)
      } catch (error) {
          res.status(500).send(error)
      }
@@ -47,12 +81,13 @@ router.get('/tasks', async(req, res) => {
 
 //------------------readById-------------------------------------------------------
  
-router.get('/tasks/:id', async(req, res) => {
+router.get('/tasks/:id', auth, async(req, res) => {
 
     const _id= req.params.id
 
           try {
-            const task = await Task.findById(_id)
+            // const task = await Task.findById(_id)
+            const task = await Task.findOne({ _id, owner: req.user._id })
             if(!task){
                 return res.status(404).send()
             }
@@ -81,7 +116,7 @@ router.get('/tasks/:id', async(req, res) => {
 
 
 //-------------patch()--------------------------------
-router.patch('/tasks/:id', async(req, res) => {
+router.patch('/tasks/:id', auth, async(req, res) => {
     const updates = Object.keys(req.body)
      const allowedToUpdate = ['description','completed']
      const isValidOperation = updates.every((update) => allowedToUpdate.includes(update))
@@ -92,7 +127,8 @@ router.patch('/tasks/:id', async(req, res) => {
 
             try {
 
-                 const task = await Task.findById(req.params.id)
+                //  const task = await Task.findById(req.params.id)
+                 const task = await Task.findBuId({_id: req.params.id, owner: req.user._id})
                  updates.forEach((update) => task[update] = req.body[update])
 
                  await task.save()
@@ -112,9 +148,10 @@ router.patch('/tasks/:id', async(req, res) => {
 
 
 //--------------delete()----------------------------------------------------
-router.delete('/tasks/:id', async(req, res) => {
+router.delete('/tasks/:id', auth, async(req, res) => {
          try {
-             const task = await Task.findByIdAndDelete(req.params.id)
+            //  const task = await Task.findByIdAndDelete(req.params.id)
+             const task = await Task.findByIdAndDelete({_id:req.params.id, owner: req.user._id})
              if(!task){
                  return res.status(404).send()
              }
